Implemented leave a review for dishes

+++ MenuUI/app/src/main/java/com/example/menuui/Review.java
@@ -2,11 +2,11 @@ package com.example.menuui;
 
 class Review {
     String review;
-    int rating;
+    float rating;
     boolean would_recommend;
     String author;
 
-    Review(String review, int rating, boolean would_recommend, String author) {
+    Review(String review, float rating, boolean would_recommend, String author) {
         this.review = review;
         this.rating = rating;
         this.would_recommend = would_recommend;
